+++ test/spec/adtech.service.ut.js
@@ -5,10 +5,13 @@
         describe('AdTechService', function() {
             var $window,
                 $q,
+                c6Defines,
                 AdTechService;
 
             beforeEach(function() {
+                c6Defines = {};
                 module('c6.rumble.services', function($provide) {
+                    $provide.value('c6Defines',c6Defines);
                     $provide.value('$window', {
                         location: {
                             hostname: 'localhost'
@@ -54,9 +57,10 @@
 
                         it('should first call ADTECH.loadAd() with the master placement id', function() {
                             expect($window.ADTECH.loadAd).toHaveBeenCalledWith({
-                                network: '5072',
+                                secure: false,
+                                network: '5473.1',
                                 server: 'adserver.adtechus.com',
-                                placement: 3214274,
+                                placement: 3220577,
                                 adContainerId: 'adtechPlacement',
                                 kv: { weburl: 'localhost' },
                                 complete: jasmine.any(Function)
@@ -75,7 +79,8 @@
 
                             it('should resolve the init() promise and call ADTECH.loadAd() again', function() {
                                 expect($window.ADTECH.loadAd).toHaveBeenCalledWith({
-                                    network: '5072',
+                                    secure: false,
+                                    network: '5473.1',
                                     server: 'adserver.adtechus.com',
                                     placement: 123456,
                                     adContainerId: 'container',
@@ -112,7 +117,8 @@
                             expect($window.ADTECH.loadAd.calls.length).toBe(loadAdCallCount + 1);
 
                             expect($window.ADTECH.loadAd.mostRecentCall.args[0]).toEqual({
-                                network: '5072',
+                                secure: false,
+                                network: '5473.1',
                                 server: 'adserver.adtechus.com',
                                 placement: 123456,
                                 adContainerId: 'container2',
@@ -131,9 +137,30 @@
                             AdTechService.loadAd({id: 'container', displayAdSource: 'waterfall'});
 
                             expect($window.ADTECH.loadAd).toHaveBeenCalledWith({
-                                network: '5072',
+                                secure: false,
+                                network: '5473.1',
                                 server: 'adserver.adtechus.com',
-                                placement: 3214274,
+                                placement: 3220577,
+                                adContainerId: 'adtechPlacement',
+                                kv: { weburl: 'parent' },
+                                complete: jasmine.any(Function)
+                            });
+                        });
+                    });
+                    
+                    describe('when protocol is https', function() {
+                        it('should set the secure property to true', function() {
+                            c6Defines.kProtocol = 'https:';
+                            $window.location.hostname = '';
+                            $window.parent.location.hostname = 'parent.com';
+
+                            AdTechService.loadAd({id: 'container', displayAdSource: 'waterfall'});
+
+                            expect($window.ADTECH.loadAd).toHaveBeenCalledWith({
+                                secure: true,
+                                network: '5473.1',
+                                server: 'adserver.adtechus.com',
+                                placement: 3220577,
                                 adContainerId: 'adtechPlacement',
                                 kv: { weburl: 'parent' },
                                 complete: jasmine.any(Function)